refactor: refactor the main entry point
